RELATED: ONE-4477 Switch enableTableColumnsGrowToFit to true
